requirements.txt
fake-useragent==2.2.0
python-dotenv==1.0.0
requests==2.32.3